requirements.txt
# Core dependencies
pydantic>=2.5.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# CLI utilities
colorama>=0.4.6

